renamed "events.js" to "hub.js"
