+++ lib/client/engine.js
@@ -1,10 +1,9 @@
-var EventEmitter = require('./events');
+var Hub = require('./hub');
 var Instance = require('./instance');
-var Transform = require('./transform');
 
 // extend engine with event emitter
-for (var key in EventEmitter) {
-    engine[key] = EventEmitter[key];
+for (var key in Hub) {
+    engine[key] = Hub[key];
 }
 engine._events = {};
 
@@ -16,7 +15,7 @@ engine.client = true;
 
 // data handlers
 engine.handlers = {
-    transform: Transform
+    transform: require('./transform')
 };
 
 Instance();
@@ -29,5 +28,5 @@ Instance();
  * @todo check for memory leaks
  */
 engine.reload = function (keepDom) {
-    engine.emit('reload', keepDom);
+    // engine.emit('reload', keepDom);
 };
