+++ lib/client/instance.js
@@ -1,4 +1,4 @@
-var EventEmitter = require('./events');
+var Hub = require('./hub');
 var Flow = require('./flow');
 var utils = require('./utils');
 
@@ -40,7 +40,7 @@ module.exports = function loadModuleInstance (name, tempLoadingCache) {
     }
 
     // create a new empty Module
-    var module_instance = utils.clone(EventEmitter);
+    var module_instance = utils.clone(Hub);
 
     // get or create an temporary loading cache array
     tempLoadingCache = tempLoadingCache || [];
